requirements.txt
# ─────────────────────────────────────────────
#  Quizzaro YouTube Automation – requirements.txt
#  All packages are free & open-source
# ─────────────────────────────────────────────

# ── AI / LLM ──────────────────────────────────
google-generativeai>=0.5.4          # Gemini API
groq>=0.9.0                         # Groq (LLaMA / Mixtral)
openai>=1.30.0                      # OpenAI (fallback)
openrouter                          # OpenRouter multi-model fallback

# ── Text-to-Speech ────────────────────────────
edge-tts>=6.1.9                     # Microsoft Edge TTS (free, high quality)
kokoro-onnx>=0.3.0                  # Kokoro TTS (open-source, high quality)

# ── Audio Processing ──────────────────────────
pydub>=0.25.1                       # Audio manipulation (speed/pitch humanizer)
soundfile>=0.12.1                   # Audio file I/O
numpy>=1.26.0                       # Numerical ops for audio
scipy>=1.13.0                       # Signal processing

# ── Video Rendering ───────────────────────────
moviepy>=1.0.3                      # Video composition engine
Pillow>=10.3.0                      # Image drawing (text, timer, watermark)
opencv-python-headless>=4.9.0.80    # Frame-level video processing
imageio>=2.34.0                     # Frame I/O
imageio-ffmpeg>=0.5.1               # FFmpeg backend for imageio

# ── HTTP / Web ────────────────────────────────
requests>=2.32.0                    # General HTTP
httpx>=0.27.0                       # Async HTTP
aiohttp>=3.9.5                      # Async HTTP (content fetching)
beautifulsoup4>=4.12.3              # HTML parsing (Wikipedia scraper)
lxml>=5.2.1                         # Fast XML/HTML parser

# ── Google / YouTube ──────────────────────────
google-api-python-client>=2.130.0   # YouTube Data API v3
google-auth>=2.29.0                 # OAuth2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# ── Data / Storage ────────────────────────────
tinydb>=4.8.0                       # Lightweight JSON database (question bank, logs)
python-dateutil>=2.9.0              # Date math (15-day / 10-day repeat checks)
pytz>=2024.1                        # Timezone handling for scheduler

# ── Utilities ─────────────────────────────────
python-dotenv>=1.0.1                # Load .env locally (secrets_loader fallback)
tenacity>=8.3.0                     # Retry logic with exponential backoff
loguru>=0.7.2                       # Structured logging
schedule>=1.2.1                     # In-process scheduler (used by bootstrap)
pytrends>=4.9.2                     # Google Trends (unofficial API)
wikipedia-api>=0.6.0                # Wikipedia content fetcher
yt-dlp>=2024.5.27                   # YouTube metadata scraper (no download)
